+++ Buffs/JadeFire.cs
@@ -4,7 +4,7 @@ using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace MokkelMod.Content.Sprites.Buffs
+namespace MokkelMod.Buffs
 {
     public class JadeFire : ModBuff
     {
