+++ Items/IlluminescentEssence.png
