+++ Items/Weapons/InstaKill.png
