+++ NPCs/AntlionEgg.png
