+++ NPCs/BroodMothera.png
