+++ NPCs/Hookcs
