+++ NPCs/Variables.cs
